Build the path correctly in entry-path.
We are traversing the tree from the inner-most subheading outward, so
we need to reverse the collected paths when building the output
directory.

Also, this function was previously moving the point, which is a huge
problem. Stop doing that so that we are only exporting one sub-heading
at a time.