Handle images in org-neuron-link
Copy code from `org-hugo-link` and keep only the bits I need. org-hugo
has a beautiful system for copying images into the static folder of
the exported dir. I am using this as is. The syntax of the output path
is slightly different, Neuron supports a much smaller subset of link
rendering styles / properties.

Only keeping support for the Markdown version of Image links and
dropping everything else.

Note that this commit introduces the `:hugo-base-dir` setting in the
derived backend, since the code for copying images uses this. It is
set to the same value as the `:neuron-base-dir`