Enable dirtree usage: Use ids in directory names.
I'm planning on using dirtree by default, and this is necessary to tie
the directory zettle to it's content zettle.

The problem with using the dirtree plugin is that the root/tags it
builds are ugly in my case. I'd like a way to turn this off, but it
does not exist in neuron (other than turning all tags off, which is
what I will do for now).

Note that this commit also cuts the ID length to the first 8 chars.
This is important because with full IDs, Emacs's auto-save starts
complaining about "filename too long" for deeply nested subtrees (3+
levels)

This commit also introduces a change for index.md file handling.
Neuron cannot handle index/ directory and index.md file. Due to this,
if the export-file-name is index, I use the ID of the entry to create
directory names. In practice, this leads to ugly directories. The way
to deal with this is to create a separate heading for the index and
avoid nesting anything under it.