Introduce DIRTREE_DISPLAY
I plan to use this flag to control whether dirtree should be displayed
on a given post or not. This requires editing the generated
front-matter, which I have been leaving to `ox-hugo` so far. I don't
plan to pick this up immediately, but committing this so that I have a
TODO to keep track of.