Refactor slug-name and file-name generation
Truncate the slug to 32 characters or less. Use '-' as a word
boundary and always truncate the slug at a word boundary, unless the
entire slug is a single large word.

Filename generation is only done if
`org-neuron-export-human-readable-filenames` is true. Otherwise, the
ZETTEL_ID is what is used everywhere.

When generating a filename, preference is: explicit filename >
explicit slug > generated filename. This filename is the one that is
used to reference zettels all over the exported content.