Refactor the collection of Neuron child posts in content
Going through this showed me that my code is entirely dependent on the
narrowing and pre-processing of content that I've already set up. The
problem with this is that tag inheritance fails since I am essentially
moving a subtree to an entirely new document and exporting it from
there whenever I export content.

I have noted this down and will fix it over time.