Respect the special case of an index filename
When EXPORT_FILE_NAME is `index`, use it over using a `zettel-id`
regardless of human-readable name settings.

This is because an index file is expected to be the default file from
which the rest of the website can be navigated. Without it, the user
will have to manually create an index file.