Control if the heading is exported in the article
The new variable `org-neuron-export-title-in-article` (OFF by default)
controls if the heading is inserted in the HTML we are generating in
our markdown. This is a very specific use-case for myself, where I
want to rely on the shape of the exported HTML instead of the YAML
metadata to determine the title of my pages.

This variable is hidden for the moment and will not affect anything in
the normal export.